+++ +gamrywb/+data/loadFilesIntoSession.m
@@ -50,7 +50,7 @@ function out = normalizeFilepaths(filepaths)
     elseif iscell(filepaths)
         out = filepaths(:).';
     else
-        error('gamrywb:ui:loadFilesIntoSession:InvalidFilepaths', ...
+        error('gamrywb:data:loadFilesIntoSession:InvalidFilepaths', ...
             'filepaths must be a char, string, or cell array.');
     end
 end